[Ref] Avoid arithmetic operations involving enums which are deprecated in C++20.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@12206 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/PatternEditorDialogs.cpp
@@ -898,7 +898,7 @@ BOOL CChordEditor::OnInitDialog()
 	ResizableDialog::OnInitDialog();
 	m_Keyboard.Init(this, (CHORD_MAX - CHORD_MIN) / 12, true);
 	// Shortcut key combo box
-	AppendNotesToControl(m_CbnShortcut, NOTE_MIN, NOTE_MIN + kcVPEndChords - kcVPStartChords);
+	AppendNotesToControl(m_CbnShortcut, NOTE_MIN, NOTE_MIN + static_cast<int>(kcVPEndChords) - static_cast<int>(kcVPStartChords));
 
 	m_CbnShortcut.SetCurSel(0);
 	// Base Note combo box
